requirements.txt
numpy==1.21.5
pandas
Flask==2.2.2
gunicorn==20.1.0
pyrebase4
firebase-admin==6.0.1
flask_restful==0.3.9
prometheus_flask_exporter==0.21.0
flask_cors